+++ tododav/__main__.py
@@ -1,5 +1,5 @@
+'''
 This module is an abstraction layer for _caldav_ and _VTODO_ objects.
 
-# Changelog
-
-The changelog [is here: CHANGELOG.md](CHANGELOG.md).
+Author: Manuel Senfft (www.tagirijus.de)
+'''